+++ test/pricing.test.js
@@ -1,18 +1,18 @@
 import { expect } from '@bundled-es-modules/chai';
-import { PricingStorageSimulator } from '../src/lib/pricing.js';
+import { PricingConsumptionSimulator } from '../src/lib/pricing.js';
 
 const INTERVALS_FOO = [
-  { minRange: 0, maxRange: 100 * 1e6, price: 0 },
-  { minRange: 100 * 1e6, maxRange: 1e12, price: 1 },
-  { minRange: 1e12, maxRange: 25 * 1e12, price: 2 },
-  { minRange: 25 * 1e12, price: 3 },
+  { minRange: 0, maxRange: 100, price: 0 },
+  { minRange: 100, maxRange: 200, price: 1 },
+  { minRange: 200, maxRange: 300, price: 2 },
+  { minRange: 300, price: 3 },
 ];
 
 const INTERVALS_BAR = [
-  { minRange: 0, maxRange: 200 * 1e6, price: 0 },
-  { minRange: 200 * 1e6, maxRange: 5 * 1e12, price: 10 },
-  { minRange: 5 * 1e12, maxRange: 50 * 1e12, price: 20 },
-  { minRange: 50 * 1e12, price: 30 },
+  { minRange: 0, maxRange: 1000, price: 0 },
+  { minRange: 100, maxRange: 2000, price: 10 },
+  { minRange: 200, maxRange: 3000, price: 20 },
+  { minRange: 3000, price: 30 },
 ];
 
 const INTERVALS = [
@@ -27,59 +27,59 @@ const NULLISH_INTERVALS = [
 
 const ONE_GIGABYTE = 1e9;
 
-describe('PricingStorageSimulator', () => {
+describe('PricingConsumptionSimulator', () => {
 
   describe('getQuantity()', () => {
 
     it('default', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
       expect(ps.getQuantity('foo')).to.equal(0);
       expect(ps.getQuantity('bar')).to.equal(0);
     });
 
     it('some value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 5 * 1e3 * ONE_GIGABYTE);
-      ps.setQuantity('bar', 10 * 1e3 * ONE_GIGABYTE);
-      expect(ps.getQuantity('foo')).to.equal(5 * 1e3 * ONE_GIGABYTE);
-      expect(ps.getQuantity('bar')).to.equal(10 * 1e3 * ONE_GIGABYTE);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 50);
+      ps.setQuantity('bar', 500);
+      expect(ps.getQuantity('foo')).to.equal(50);
+      expect(ps.getQuantity('bar')).to.equal(500);
     });
   });
 
   describe('getCurrentInterval()', () => {
 
     it('default', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
       expect(ps.getCurrentInterval('foo')).to.equal(INTERVALS_FOO[0]);
       expect(ps.getCurrentInterval('bar')).to.equal(INTERVALS_BAR[0]);
     });
 
     it('inside interval', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 150 * 1e6);
-      ps.setQuantity('bar', 250 * 1e6);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 150);
+      ps.setQuantity('bar', 1500);
       expect(ps.getCurrentInterval('foo')).to.equal(INTERVALS_FOO[1]);
       expect(ps.getCurrentInterval('bar')).to.equal(INTERVALS_BAR[1]);
     });
 
     it('lower limit of interval', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 1e12);
-      ps.setQuantity('bar', 5 * 1e12);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 200);
+      ps.setQuantity('bar', 2000);
       expect(ps.getCurrentInterval('foo')).to.equal(INTERVALS_FOO[2]);
       expect(ps.getCurrentInterval('bar')).to.equal(INTERVALS_BAR[2]);
     });
 
     it('big value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 50 * 1e12);
-      ps.setQuantity('bar', 100 * 1e12);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 1000);
+      ps.setQuantity('bar', 10000);
       expect(ps.getCurrentInterval('foo')).to.equal(INTERVALS_FOO[3]);
       expect(ps.getCurrentInterval('bar')).to.equal(INTERVALS_BAR[3]);
     });
 
     it('nullish intervals', () => {
-      const ps = new PricingStorageSimulator(NULLISH_INTERVALS);
+      const ps = new PricingConsumptionSimulator(NULLISH_INTERVALS);
       expect(ps.getCurrentInterval('foo')).to.equal(null);
       expect(ps.getCurrentInterval('bar')).to.equal(null);
     });
@@ -88,23 +88,23 @@ describe('PricingStorageSimulator', () => {
   describe('getEstimatedPrice()', () => {
 
     it('small value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 50 * 1e6);
-      ps.setQuantity('bar', 150 * 1e6);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 50);
+      ps.setQuantity('bar', 500);
       expect(ps.getEstimatedPrice('foo')).to.equal(0);
       expect(ps.getEstimatedPrice('bar')).to.equal(0);
     });
 
     it('big value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 5 * 1e3 * ONE_GIGABYTE);
-      ps.setQuantity('bar', 10 * 1e3 * ONE_GIGABYTE);
-      expect(ps.getEstimatedPrice('foo')).to.equal(5 * 1e3 * 2);
-      expect(ps.getEstimatedPrice('bar')).to.equal(10 * 1e3 * 20);
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 250);
+      ps.setQuantity('bar', 5000);
+      expect(ps.getEstimatedPrice('foo')).to.equal(250 * 2);
+      expect(ps.getEstimatedPrice('bar')).to.equal(5000 * 30);
     });
 
     it('nullish intervals', () => {
-      const ps = new PricingStorageSimulator(NULLISH_INTERVALS);
+      const ps = new PricingConsumptionSimulator(NULLISH_INTERVALS);
       expect(ps.getEstimatedPrice('foo')).to.equal(0);
       expect(ps.getEstimatedPrice('bar')).to.equal(0);
     });
@@ -113,21 +113,21 @@ describe('PricingStorageSimulator', () => {
   describe('getTotalPrice()', () => {
 
     it('small value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 50 * 1e6);
-      ps.setQuantity('bar', 150 * 1e6);
-      expect(ps.getTotalPrice()).to.equal((50 * 1e6 * 0) + (150 * 1e6 * 0));
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 50);
+      ps.setQuantity('bar', 500);
+      expect(ps.getTotalPrice()).to.equal((50 * 0) + (500 * 0));
     });
 
     it('big value', () => {
-      const ps = new PricingStorageSimulator(INTERVALS);
-      ps.setQuantity('foo', 5 * 1e3 * ONE_GIGABYTE);
-      ps.setQuantity('bar', 10 * 1e3 * ONE_GIGABYTE);
-      expect(ps.getTotalPrice()).to.equal((5 * 1e3 * 2) + (10 * 1e3 * 20));
+      const ps = new PricingConsumptionSimulator(INTERVALS);
+      ps.setQuantity('foo', 250);
+      ps.setQuantity('bar', 5000);
+      expect(ps.getTotalPrice()).to.equal((250 * 2) + (5000 * 30));
     });
 
     it('nullish intervals', () => {
-      const ps = new PricingStorageSimulator(NULLISH_INTERVALS);
+      const ps = new PricingConsumptionSimulator(NULLISH_INTERVALS);
       expect(ps.getTotalPrice('foo')).to.equal(0);
     });
   });